BaseGame

Description

This contains functions for a simple game.

Documentation

doNothing :: Step

fromStep :: Step -> Game

if'' :: Bool -> (a -> a) -> a -> a

(%) :: String -> String -> String

(%%) :: String -> Card -> String

illegal :: Int -> String -> Game

Returns a game where the move has not occured and a penalty has been given. Should be used with doOnly

penalty :: Int -> String -> Name -> Step

A penalty which does not end the turn

broadcastp :: Name -> String -> Step

Broadcast a message while identifying the player it came from.

sayAct :: Game

Broadcast the message a player sent with their action.

addPlayer :: Name -> Maybe (Name, Name) -> Step

Add a player to the game

addToSeat :: Name -> Maybe (Name, Name) -> [Name] -> [Name]

play :: Name -> Event -> Card -> Step

Play a move and tidy up the game state.

baseAct :: Game

The default behaviour. Draws always succeed, a card may be played if it is in the player's hand, it is the player's turn and it matches either the suit or rank of the previous card. Timeouts give one penalty card to the player whose turn it is.

broadcast :: String -> Step

Add a message to the list of messages. Will be seen by all players by default.

draw :: Int -> Name -> Step

Move n cards from the deck to a player's hand

getHand :: Name -> GameState -> Maybe Hand

Get the hand of the current player.

draw1 :: Name -> Step

Move a single card from the deck to a player's hand.

taxes :: Step

If any players have more than their share of cards, give return them to the deck (so that players can always draw cards and be given penalties).

cardFromDeck :: GameState -> (Card, GameState)

Remove the top card of the deck shuffling the pile back in and applying taxes if necessary.

cardToPile :: Card -> Step

Put a card on the pile

cardFromHand' :: Name -> Card -> Step

Remove a card from a players hand.

isTurn :: Name -> GameState -> Bool

Test if it is a player's turn.

nextTurn :: Step

Advance the turn order by one.

win :: Name -> Step

incomplete?