DataTypes

Description

Datatypes for cards, game state and game view alongside some basic functions to process them.

Basic Rules

type Step = GameState -> GameState

The type of transformations of game state

type Game = Event -> Step

This is the type of games (meaning a ruleset, rather than a specific instance of a game). Given an event such as a player action, and a game state, it returns the resulting state

type Rule = Game -> Game

The type of rules - a rule is added by function composition forming a chain such as 'r3 (r1 (r2 baseAct))' . Most rules should call the first argument under most circumstances and usually without changing the event.

Game -> Game == (Event -> GameState -> GameState) -> Event -> GameState -> GameState

Structures

type Name = String

Names of players.

data Action

At any point, a player can attempt to take one of these actions

Constructors

Draw Int

Play Card

data Event

Events that rules should be able to deal with

Constructors

PlayerJoin Name (Maybe (Name, Name))
A player requesting to join the game. This is not limited to the start

Action Name Action String
A player performing one of the above actions and sending a message.

Timeout
If no player has made an action for 10 seconds, this event will be sent

data GameState

The state of a game in play

Constructors

GS

Fields

_players :: [Name]

The players curretly in the game. The player whose turn it is should be at head of list and usually this advances forward by 1 each turn.

_hands :: Map Name Hand

Stores the contents of each player's hand

_deck :: [Card]

The deck from which cards are drawn

_pile :: NonEmpty Card

The cards that have been played -

_messages :: [String]

The messages that . Contains only those generated by the most recent event.

_lastMoveLegal :: Bool

Indicates if the last move was successful. This should be true When a new event happens, this is False until baseAct is called. _prevGS :: Maybe (GameState,Action),

_randg :: StdGen

A seeded random number generator so that you can

_winner :: Maybe Name

Nothing until a player p wins at which point it becomes Just p

_varMap :: Map VarName Int

A store of named variables that rules may use to keep track of state between events.

newGame :: [String] -> GameState

Construct a new game from a list of player names.

Cards

type Hand = [Card]

An ordered collection of cards forming a player's hand

type Card = (Rank, Suit)

A single playing card

data Suit

Constructors

Clubs

Diamonds

Hearts

Spades

data Rank

Constructors

Ace

Two

Three

Four

Five

Six

Seven

Eight

Nine

Ten

Jack

Knight

Queen

King

suit :: Card -> Suit

Get the suit of a card.

rank :: Card -> Rank

Get the rank of a card.

Variables

type VarName = String

setVar :: VarName -> Int -> Step

readVar :: VarName -> GameState -> Int

variable processing

modifyVar :: VarName -> (Int -> Int) -> Step

other functions

shuffleDeck :: Step

Shuffle the deck (does not touch the pile or hands) using the random seed contained in GameState.

eventPlayer :: Event -> Maybe Name

Get the playeer if an event

View Rules

type Viewer = Name -> GameState -> GameView

Functions to tell a player what they should see

type ViewRule = Viewer -> Viewer

Rules that modify what players see without affecting the game state

data CardView

A card as viewed - the type of cards sent to the client

Constructors

CardFace Card

CardBack

Structure

data GameView

The structure describing data seen by players

Constructors

GV

Fields

_handsV :: [(Name, [CardView])]

The players and the information a particular player will have about each hand. By default, this is a CardFaces for the viewing player and a number of CardBacks for others.

_pileV :: [CardView]

What should be seen of the pile. By default, only the top card is visible.

_deckV :: [CardView]

What should be seen of the pile. By default, none are visible.

_messagesV :: [String]

The messages a player can see. By default, this is all messages that have been sent

Combined Rules

type Rule' = (Rule, ViewRule)

Complex rules

Cards

Printing

uniCard :: Card -> Char

Get the unicode playing card character corresponding to some card

Parsing

type Parser = StateT String Maybe

runParser :: Parser a -> String -> Maybe a

parseSuit :: Parser Suit

parseRank :: Parser Rank

parseCard :: Parser Card

Angus needs to get rid of these

if' :: Bool -> a -> a -> a

(/\) infixl 7

Arguments

:: Functor f

=> ((a -> (a, a)) -> c -> (a, c))
Lens' c a

-> ((b -> (b, b)) -> c -> (b, c))
Lens' c b

-> ((a, b) -> f (a, b)) -> c -> f c
Lens' c (a, b)

lenses

players :: Lens' GameState [Name]

hands :: Lens' GameState (Map Name Hand)

deck :: Lens' GameState [Card]

pile :: Lens' GameState (NonEmpty Card)

messages :: Lens' GameState [String]

lastMoveLegal :: Lens' GameState Bool

randg :: Lens' GameState StdGen

winner :: Lens' GameState (Maybe Name)

varMap :: Lens' GameState (Map VarName Int)

handsV :: Lens' GameView [(Name, [CardView])]

pileV :: Lens' GameView [CardView]

deckV :: Lens' GameView [CardView]

messagesV :: Lens' GameView [String]

Orphan instances

(Enum a, Enum b, Bounded a, Bounded b, Eq a, Eq b) => Enum (a, b)

Methods

succ :: (a, b) -> (a, b)

pred :: (a, b) -> (a, b)

toEnum :: Int -> (a, b)

fromEnum :: (a, b) -> Int

enumFrom :: (a, b) -> [(a, b)]

enumFromThen :: (a, b) -> (a, b) -> [(a, b)]

enumFromTo :: (a, b) -> (a, b) -> [(a, b)]

enumFromThenTo :: (a, b) -> (a, b) -> (a, b) -> [(a, b)]