Game.Bartok.Sample

Description

Look at the implementation of these to see examples of how to write rules

Documentation

r8 :: Rule

When an 8 is played, skip the following player's turn

reverseDirection :: Step

rq :: Rule

Reverse direction when a Queen is played

mustdo7 :: Int -> (Bool -> Game) -> Rule

count7s :: GameState -> Int

r7 :: Rule

When a seven is played, the player must say 'have a nice day'. The following player must either play another seven, saying 'have a very nice day' or say 'thank you' and draw 2 cards. If futher sevens are played (each such player must add another very), the first player who cannot play one must say 'thank you very{n} much' and draw 2n cards.

r7' :: Rule

rC :: Game -> Game

gSnap :: Rule'

nextTurnDo :: String -> (Game -> Event -> GameState -> Bool) -> Rule -> Rule

nextTurnDoUntil :: String -> (Game -> Event -> GameState -> Bool) -> (Game -> Event -> GameState -> (GameState, Bool)) -> Rule

r6 :: Rule

rPM :: Rule

rPMV :: ViewRule

rPM' :: Rule'

rBlind :: Rule

rBlindV :: ViewRule

rBlind' :: Rule'

run7 :: Int -> Rule

rKnight :: Rule

rbadger :: (Event -> GameState -> GameState) -> Event -> GameState -> GameState

rFourShow :: Rule'

When a four is played, toggle complete visibility