Game.Bartok.Tutorial

Description

Documentation

Note that this file involves copy-pasting a bunch of definitions so that it generates nice haddock. It may be inaccurate or out of date.

Framework

type Game = Event -> GameState -> GameState

A Game is essentially a ruleset that defines how an event modifies the gamestate.

For example, in the following game nothing that happens has any effect:

useless :: Game
useless event state = state

data Event

A game of Bartok, from the server's point of view, consists of events arriving and affecting the gamestate.

Actions are the most interesting, and always come with a message (which may be the empty string).

Constructors

Action Name Action String
A player actually doing something (playing a card or drawing cards)

Timeout
If no player has made an action for 10 seconds, this event will be sent.

PlayerJoin Name (Maybe (Name, Name))
This is handled by the base game. You probably want to leave it alone.

PlayerLeave Name
A player leaving the game. Again handled by the base game.

type Rule = Game -> Game

A game is a composition of rules such as 'r1 (r2 (r3 useless))' .

For example, the following rule will stop players being penalized for timeouts.

rLongTurns :: Rule
rLongTurns oldgame Timeout gamestate = gamestate
rLongTurns oldgame event gamestate = oldgame event gamestate

You usually want to call oldgame event gamestate at some point so that rules which have been implemented before yours get a chance to do something.

The rest of this tutorial is unfinished, and DataTypes is a better reference. See TLib for a framework to make rules and the source code of TSample for sample rules using it.

Gamestate

type Step = GameState -> GameState

Transformations of gamestate.

type Card = (Rank, Suit)

A single playing card

type Hand = [Card]

An ordered collection of cards forming a player's hand

type Name = String

Names of players.

data Action

At any point, a player can attempt to take one of these actions

Constructors

Draw Int
Draw some number of cards

Play Card
Play a card

data GameState

The state of a game in play

Constructors

GS

Fields

_players :: [Name]

The players curretly in the game. The player whose turn it is should be at head of list and usually this advances forward by 1 each turn.

_hands :: Map Name Hand

Stores the contents of each player's hand

_deck :: [Card]

The deck from which cards are drawn

_pile :: NonEmpty Card

The cards that have been played - shuffled back into the deck when necessary.

_messages :: [String]

The messages that . Contains only those generated by the most recent event.

_lastMoveLegal :: Bool

Indicates if the last move was successful. When a new event happens, this is False until baseAct is called.

_randg :: StdGen

A seeded random number generator

_winner :: Maybe Name

Nothing until a player p wins at which point it becomes Just p

_varMap :: Map VarName Int

A store of named variables that rules may use to keep track of state between events.

View Rules

data CardView

A card as viewed - the type of cards sent to the client

Constructors

CardFace Card

CardBack

data GameView

The structure describing data seen by players

Constructors

GV

Fields

_handsV :: [(Name, [CardView])]

The players and the information a particular player will have about each hand. By default, this is a CardFaces for the viewing player and a number of CardBacks for others.

_pileV :: [CardView]

What should be seen of the pile. By default, only the top card is visible.

_deckV :: [CardView]

What should be seen of the pile. By default, none are visible.

_messagesV :: [String]

The messages a player can see. By default, this is all messages that have been sent

type Viewer = Name -> GameState -> GameView

Functions to tell a player what they should see

type ViewRule = Viewer -> Viewer

Rules that modify what players see without affecting the game state

type Rule' = (Rule, ViewRule)

Complex rules

suit :: Card -> Suit

Get the suit of a card.

rank :: Card -> Rank

Get the rank of a card.

rankChar :: Rank -> Char

One character corresponding to the rank (A1..9TJCQK)

uniCard :: Card -> Char

Get the unicode playing card character corresponding to some card

(<|>) :: Parser a -> Parser a -> Parser a

Given 2 parsers, tries the first, if it fails, try the second

eventPlayer :: Event -> Maybe Name

Get the playeer if an event

readVar :: VarName -> GameState -> Int

variable processing

shuffleDeck :: Step

Shuffle the deck (does not touch the pile or hands) using the random seed contained in GameState.

newGame :: [String] -> GameState

Construct a new game from a list of player names.

Orphan instances

(Enum a, Enum b, Bounded a, Bounded b, Eq a, Eq b) => Enum (a, b)

Instance details

Methods

succ :: (a, b) -> (a, b)

pred :: (a, b) -> (a, b)

toEnum :: Int -> (a, b)

fromEnum :: (a, b) -> Int

enumFrom :: (a, b) -> [(a, b)]

enumFromThen :: (a, b) -> (a, b) -> [(a, b)]

enumFromTo :: (a, b) -> (a, b) -> [(a, b)]

enumFromThenTo :: (a, b) -> (a, b) -> (a, b) -> [(a, b)]