RuleHelpers

Documentation

when :: (a -> Bool) -> Rule -> a -> Rule

with :: (Event -> GameState -> a) -> (a -> Rule) -> Rule

with' :: (Event -> GameState -> a) -> (a -> Game) -> Game

when' :: (a -> Bool) -> Game -> a -> Game

doAfter :: Game -> Rule

doBefore :: Game -> Rule

doOnly :: Game -> Rule

regexProcess :: String -> Regex

player's next action must be the given one how do I make require actions for something other than a single

split :: String -> [String]

reconstitute :: [String] -> String

findInMs :: String -> String -> Bool

removeIn :: String -> String -> (Bool, String)

removeIn' :: String -> String -> (Maybe String, String)

removeIn'' :: Regex -> [String] -> [String] -> (Maybe String, [String], [String])

removeAll :: String -> String -> ([String], String)

removeAllN :: String -> String -> (Int, String)

removeAll' :: Regex -> ([String], [String]) -> [String] -> ([String], [String])

mustSay :: String -> Game

mustSay' :: String -> Event -> (Event, Step)

said :: String -> String -> Bool

banPhrase :: Int -> String -> ((Name, Action, String) -> GameState -> Bool) -> Rule

require :: (Name, Action, String) -> (Bool -> Game) -> Rule

onPlay :: (Card -> Rule) -> Rule

onLegalCard :: (Card -> Game) -> Rule

onAction :: ((Name, Action, String) -> Rule) -> Rule

onDraw :: ((Name, Int) -> Rule) -> Rule

nextTurn :: Step

Advance the turn order by one.

doNothing :: Step

illegal :: Int -> String -> Game

Returns a game where the move has not occured and a penalty has been given. Should be used with doOnly

broadcast :: String -> Step

Add a message to the list of messages. Will be seen by all players by default.

broadcastp :: Name -> String -> Step

Broadcast a message while identifying the player it came from.

penalty :: Int -> String -> Name -> Step

A penalty which does not end the turn

isTurn :: Name -> GameState -> Bool

Test if it is a player's turn.

win :: Name -> Step

incomplete?

draw :: Int -> Name -> Step

Move n cards from the deck to a player's hand