TLib

Description

This presents operations with more arguments than strictly necessary so that they can be used interchangeably

The fundamental type

type GEGSto a

Arguments

= Game
The inner ruleset

-> Event
The event under consideration

-> GameState
The previous state of the game

-> a

Avoids having to write out extremely long types. Almost a Monad.

users

withAction :: (Action -> Rule) -> Rule

When an action happens, do something with it

withMessage :: (String -> Rule) -> Rule

When an action happens, do something with the message that gets sent

withCard :: (Card -> Rule) -> Rule

When an card is played happens, do something with it

with

Arguments

:: GEGSto a
The extractor

-> (a -> GEGSto b)
The thing to do with the value

-> GEGSto b

Build a rule (or similar that depends on an extracted value (if you know monads, this is bind (>>=)).

e.g. with (getVar "n") ( n -> doSomethingInvolvingN)

e.g. > with (getVar "n") ( n -> doSomethingInvolvingN)

Operators

when :: GEGSto Bool -> Rule -> Rule

Given a condition, implement a rule only when that condition holds.

whether

Arguments

:: GEGSto Bool
if

-> Rule
then

-> Rule
else

-> Rule

Given a condition, do one thing when the condition holds and another when it doesn't.

onNextTurn

Arguments

:: VarName
a fresh varName

-> GEGSto Bool
the condition which must be met

-> Rule
what to do on the next turn

-> Rule

Given a fresh variable name, when a condition is met, run an action on the following turn.

uponDoUntil :: VarName -> GEGSto Bool -> Rule -> GEGSto Bool -> Rule

Upon a condition being satisfied, implement a rule until a second condition is satisfied.

(.:) :: (a -> b) -> GEGSto a -> GEGSto b

lift an action (monadic fmap)

__ :: a -> GEGSto a

Abbreviation for _ _ _ ->

Tests

isAction :: GEGSto Bool

Returns True if the event is an Action.

isLegal :: GEGSto Bool

Returns True if the event is an action and is legal under the inner ruleset.

isTurn :: GEGSto Bool

Returns True if the event is an action and the next Turn should be taken by the player submitting it.

actionIs :: (Action -> Bool) -> GEGSto Bool

Returns True if the event is an action which satisfies the condition.

cardIs

Arguments

:: (Card -> Bool)
The condition

-> GEGSto Bool

Returns True if the event is a Play action and the card satisfies the condition.

boolVar :: VarName -> GEGSto Bool

Returns True if the value of the named variable is not 0.

said :: String -> GEGSto Bool

Tests if a player said something matching a given regex (case insenstive). Like most string processing functions in this project, this separates input messages based on semicolons.

(~&~) :: GEGSto Bool -> GEGSto Bool -> GEGSto Bool

Combines 2 tests. Returns True if both do.

not_ :: GEGSto Bool -> GEGSto Bool

Negates a test. Be aware that (not_ (actionIs (==(Draw 2)))) is different from (actionIs (/=(Draw 2))). The first example would return True for events that are not actions.

getters

getVar :: String -> GEGSto Int

Retrive a variable from the GameState.

state :: GEGSto GameState

Extract the state argument.

Helper Functions

mkFailMsg :: String -> String

The penalty message produced by mustSay s

mustSay :: String -> Rule

penalize if s is not said

unnec :: String -> Rule

Penalize for unnecessarily saying something. This works by testing if the absence of any matching component of a player's message would cause a penalty containing failmsg. If not, the component is assumed to be unnecessary.

sometimesSay :: String -> GEGSto Bool -> Rule

If the condition holds, require that they say the message; otherwise, penalize them if they say it

mustDo :: Action -> Rule -> Rule

Any action except the specified one is an illegal move. When the specified action occurs, run the second argument.

mustSayPenalty :: String -> String -> Rule

Like mustSay, but with custom penalty message

sometimesSayPenalty :: String -> GEGSto Bool -> String -> String -> Rule

unnecPenalty :: String -> (String -> String) -> Rule

modifyPlayers :: ([Name] -> [Name]) -> Step

Transform the list of players (affects turn order).

class

class Ruleable a where

A class for types that are smaller than Rule so that there are different options for how to convert.

Methods

doAfter

Arguments

:: a

-> Rule
Perform an action after the inner ruleset.

doBefore

Arguments

:: a

-> Rule
Perform an action before the inner ruleset.

doOnly

Arguments

:: a

-> Rule
Perform an action and ignore the inner ruleset. This should be used rarely.

From DataTypes:

type Step = GameState -> GameState

The type of transformations of game state

type Game = Event -> Step

This is the type of games (meaning a ruleset, rather than a specific instance of a game). Given an event such as a player action, and a game state, it returns the resulting state

type Rule = Game -> Game

The type of rules - a rule is added by function composition forming a chain such as 'r3 (r1 (r2 baseAct))' . Most rules should call the first argument under most circumstances and usually without changing the event.

Game -> Game == (Event -> GameState -> GameState) -> Event -> GameState -> GameState

data GameState

The state of a game in play

data Event

Events that rules should be able to deal with

Constructors

PlayerLeave Name
A player leaving the game

PlayerJoin Name (Maybe (Name, Name))
A player requesting to join the game. This is not limited to the start

Action Name Action String
A player performing one of the above actions and sending a message.

Timeout
If no player has made an action for 10 seconds, this event will be sent

data Action

At any point, a player can attempt to take one of these actions

Constructors

Draw Int

Play Card

type Name = String

Names of players.

cards

type Card = (Rank, Suit)

A single playing card

data CardView

A card as viewed - the type of cards sent to the client

Constructors

CardFace Card

CardBack

data GameView

The structure describing data seen by players

Constructors

GV

Fields

_handsV :: [(Name, [CardView])]

The players and the information a particular player will have about each hand. By default, this is a CardFaces for the viewing player and a number of CardBacks for others.

_pileV :: [CardView]

What should be seen of the pile. By default, only the top card is visible.

_deckV :: [CardView]

What should be seen of the pile. By default, none are visible.

_messagesV :: [String]

The messages a player can see. By default, this is all messages that have been sent

suit :: Card -> Suit

Get the suit of a card.

rank :: Card -> Rank

Get the rank of a card.

data Suit

Constructors

Clubs

Diamonds

Hearts

Spades

data Rank

Constructors

Ace

Two

Three

Four

Five

Six

Seven

Eight

Nine

Ten

Jack

Knight

Queen

King

type Viewer = Name -> GameState -> GameView

Functions to tell a player what they should see

type ViewRule = Viewer -> Viewer

Rules that modify what players see without affecting the game state

readVar :: VarName -> GameState -> Int

variable processing

setVar :: VarName -> Int -> Step

modifyVar :: VarName -> (Int -> Int) -> Step

nextTurn :: Step

Advance the turn order by one.

draw :: Int -> Name -> Step

Move n cards from the deck to a player's hand

broadcast :: String -> Step

Add a message to the list of messages. Will be seen by all players by default.

(%) :: String -> String -> String

mapAllCards :: (Card -> Card) -> GameState -> GameState

Modify all cards in the game