Views

Description

The functions in this module define how the data sent to clients is constructed. They should help when writing rules that affect what players can see.

Type Synonyms

type Viewer = Name -> GameState -> GameView

Functions to tell a player what they should see

type ViewRule = Viewer -> Viewer

Rules that modify what players see without affecting the game state

Basic Rules

baseViewer :: Viewer

Like baseGame, this defines how cards should be viewed in a very simple game. See GameView for more.

data GameView

The structure describing data seen by players

Constructors

GV

Fields

_handsV :: [(Name, [CardView])]

The players and the information a particular player will have about each hand. By default, this is a CardFaces for the viewing player and a number of CardBacks for others.

_pileV :: [CardView]

What should be seen of the pile. By default, only the top card is visible.

_deckV :: [CardView]

What should be seen of the pile. By default, none are visible.

_messagesV :: [String]

The messages a player can see. By default, this is all messages that have been sent

lastCard :: ViewRule

Hide the last card in a player's hand.

onlyOwnCards :: ViewRule

Hides other players' hands.

defaultView :: Viewer

baseView with the above rules added to it.

Helper functions

dealing with hands

mapOwnHand :: ([CardView] -> [CardView]) -> ViewRule

Modify a player's view of their own hand

mapOtherHands :: ([CardView] -> [CardView]) -> ViewRule

Modify the view of all other players' hands

modifyHandMap :: ([(Name, [CardView])] -> [(Name, [CardView])]) -> ViewRule

Apply a modification to the structure describing what can be seen of each player's hand after inner rules have been applied.

mapHands :: ([CardView] -> [CardView]) -> ViewRule

Apply a function to all views of hands

hideHand :: [CardView] -> [CardView]